+++ src/main/java/com/academy/mudogroupware/file/application/usecase/GetFileDownloadUrlUseCase.java
@@ -1,0 +1,6 @@
+package com.academy.mudogroupware.file.application.usecase;
+
+public interface GetFileDownloadUrlUseCase {
+
+    String getDownloadUrl(Long fileId);
+}