+++ src/main/java/com/academy/mudogroupware/notice/application/command/NoticeAttachmentInput.java
@@ -1,8 +1,7 @@
 package com.academy.mudogroupware.notice.application.command;
 
 public record NoticeAttachmentInput(
-        String fileUrl,
-        String fileName,
-        String fileType
+        Long fileId,
+        String fileName
 ) {
 }