+++ src/main/java/com/academy/mudogroupware/notice/application/service/CreateNoticeService.java
@@ -42,6 +42,6 @@ public class CreateNoticeService implements CreateNoticeUseCase {
     }
 
     private NoticeAttachment toAttachment(NoticeAttachmentInput input) {
-        return NoticeAttachment.create(input.fileUrl(), input.fileName(), input.fileType());
+        return NoticeAttachment.create(input.fileId(), input.fileName());
     }
 }